Adapt for python3; add pandas reader
Removing logging commands greatly increases read speed
